feat: Finalizing route documentation with Swagger and adding unit test files.

+++ src/controllers/rolesPermissions.controllers.ts
@@ -2,12 +2,12 @@ import { FastifyRequest, FastifyReply } from "fastify";
 import {
   roleIdSchema,
   rolePermissionSchema,
-} from '../schemas/rolesPermissions.schemas.js';
+} from "../schemas/rolesPermissions.schemas.js";
 import {
   createRolePermissionsService,
   deleteRolePermissionsService,
   getRolePermissionsService,
-} from '../services/rolesPermissions.services.js';
+} from "../services/rolesPermissions.services.js";
 
 export async function getRolePermissionsController(
   request: FastifyRequest,
@@ -18,7 +18,9 @@ export async function getRolePermissionsController(
     const roles = await getRolePermissionsService(id);
 
     if (roles.length === 0) {
-      throw new Error("Nenhuma permissão encontrada para esse cargo");
+      return reply
+        .status(404)
+        .send({ message: "Nenhuma permissão encontrada para esse cargo" });
     }
 
     return reply.status(200).send(roles);
@@ -32,9 +34,23 @@ export async function createRolePermissionsController(
   reply: FastifyReply
 ) {
   try {
-    const parsedData = rolePermissionSchema.parse(request.body);
-    const rolePermission = await createRolePermissionsService(parsedData);
-    return reply.status(201).send(rolePermission);
+    const { role_id, permission_id } = rolePermissionSchema.parse(request.body);
+
+    const permissions = await getRolePermissionsService(role_id);
+
+    if (permissions.some((item) => permission_id.includes(item.permission_id))) {
+      return reply
+        .status(400)
+        .send({ message: "O cargo já possui uma das permissões mencionadas." });
+    }
+
+    const rolePermission = await createRolePermissionsService({
+      role_id,
+      permission_id,
+    });
+    return reply
+      .status(201)
+      .send({ message: "Permissões liberadas com sucesso!" });
   } catch (error: any) {
     return reply.status(400).send({ message: error.message });
   }
@@ -45,12 +61,24 @@ export async function deleteRolePermissionController(
   reply: FastifyReply
 ) {
   try {
-    const parsedData = rolePermissionSchema.parse(request.body);
-    const deletedRolePermission = await deleteRolePermissionsService(
-      parsedData
-    );
+    const { role_id, permission_id } = rolePermissionSchema.parse(request.body);
 
-    return reply.status(200).send(deletedRolePermission);
+    const permissions = await getRolePermissionsService(role_id);
+
+    if (!permissions.some((item) => permission_id.includes(item.permission_id))) {
+      return reply
+        .status(400)
+        .send({ message: "O cargo não possui alguma das permissões mencionadas." });
+    }
+
+    const deletedRolePermission = await deleteRolePermissionsService({
+      role_id,
+      permission_id,
+    });
+
+    return reply
+      .status(200)
+      .send({ message: "Permissões removidas com sucesso!" });
   } catch (error: any) {
     return reply.status(400).send({ message: error.message });
   }
